Implement APU and basic sound
I took my first shot at implementing the APU,
for now just implementing the triangle channel.
This also required adding sound output via SDL.

The output is admittedly a huge mess at the moment,
and has a lot of problems, but since it's not
currently ear-destroying, I think it's good enough
for my first APU checkin.

The APU pushes to a sound buffer every APU clock,
and during each frame, the buffered sound is
downsampled via push_audio() and stored in a second
buffer. Finally, SDL is configured so that the audio
device will call audio_callback, which reads the
next chunk of sound and pushes it to the audio device.
The audio buffer is treated like a ring buffer, so
the writer will loop around till it bumps into the
reader, and the reader will read until it bumps into
the writer, at which point it will output silence
until more audio is supplied.

This probably sounds convoluted, but it's the best
setup I could figure out. There's a lot of work
to be done on the audio output processing end,
but I'll figure it out eventually. This is at
least a stable implementation that does something
resembling working right.