Fix issues with sprite rendering
The PPU now skips empty sprite slots. It's not clear how the real NES handles
this, but I found it easiest to just keep a count of the number of sprites.
In addition, I also implemented sprite disable and sprite priority (which
enables sprites to be drawn behind the background).