Upgrade compatibility list to web page
To make things a bit nicer and easier to keep track of, the compatibility list
is now a web page, with every game from the officially-licensed NES library
included. The one fancy feature is that the percentage for each group is
automatically calculated. Other than that, it's very very basic. But hey,
I'm not a front-end developer. Might improve it later.

Also might set up proper hosting? I'll have to look into how that works.