Restructure code for movie playback feature
This commit does a lot of restructuring of the code, mainly pulling out stuff
into separate functions, since the TAS playback module and the regular
play module are both separate and need to utilize the emulator's functions
differently. It's not quite there yet, but since this is such a major change,
I wanted to get it checked in anyway. Right now, movies appear to play back,
but they seem to do so inaccurately for reasons that aren't fully clear yet.

build_nes.bat is updated with the changes, but the makefile will be broken?