Fix movie playback bug
It turns out I was feeding the emulator inputs a frame too late,
causing the SMB TAS to play back incorrectly. This should be fixed
now.

Also smoothed out a number of minor things, mostly relating to zeroing
memory on init and setting variables to be unsigned when there's no
reason for them to be signed.