Increase TAS playback accuracy
There were a number of issues causing frame desync while trying to play
back a TAS movie. Proper disabling of the NMI_occurred flag on PPUSTATUS
reads has been implemented. There were some inaccuracies in cycle counts,
in particular the check for branches crossing page boundaries was wrong.

Unfortunately, while the SMB playback is now highly accurate, it's not
frame-perfect yet. In FCEUX, the init routine seems to busy wait on vblank,
but miss a frame, probably due to reading on the same frame vblank is set.
Emulating this would require cycle-perfect accuracy. It's the simple truth
that frame-perfect accuracy is more fine-grained than we'd like to think.
The fix for this right now is to edit the movie by removing one blank input
near the start. With this tweak, it plays back perfectly.