Add sound post-processing and movie speedup key
Did some work interpolating between sound and silence to get less popping.
A small change to sample calculation also seemed to vastly improve the clarity
of the sound output. It's not perfect, but this is a huge improvement.

Also the addition of a speedup key for movies (+) is in there too. Should
probably have been a separate commit, but these are both in the same file,
so...