Implement pulse channel 1
First pulse channel should now be working, with all the nice features like
sweep and envelope working properly. Also removed a lot of the fiddly
post-processing, because something just seemed to fix the popping, probably
reducing the volume of the triangle channel. With multiple channels, a
post-processing algorithm would need to be more sophisticated to account
for this stuff anyway.

Pulse channel 2 ought to be a near carbon copy, but I wanted to get this
checked in before piling on even more stuff.