Implement UNROM and new mapper structure
The cartridge code is now designed to hold function pointers to the underlying
mapper functions, rather than do anything itself. This will allow the
progressive installation of new mappers without any need to mess with
outside code, other than slotting the mapper functions into the cartridge
table.

The first candidate for this new structure is of course UNROM, which is
not altogether very different from NROM, other than a bit of PRG
bank-switching. This will help to expand the pool of working games,
although I think the focus next should be overhauling the memory-access
code to not rely on the awful notify/accessed pattern I've got now.