Refactor memory access code
The way memory accesses worked previously is that I would just grab a pointer
to the appropriate bit of memory and modify it. This was great for a while,
by which I mean it almost immediately caused problems. Now I'm passing in
a pointer to the data byte that I'm writing/reading to the address, so that
any sort of necessary side effects can happen without ugly, ugly hacks.

A whole raft of functions need to be renamed now in line with their new
functionality (i.e., 'access_foo_memory', not 'get_pointer_at_foo_address'),
and there are now unnecessary memory access state-tracking variables that
should be removed, but I'm happy enough that this whole conversion worked
more or less without issues. So I'll take care of that next push.