Fix sprite 0 hit check
I was incorrectly only flagging sprite 0 hits if the sprite priority was
set to 'background'. As it happens, this solves the problem with Bomberman.

Unfortunately, Super Mario Bros must be downgraded to 'runs' until the
input issue is solved. It might seem a bit silly that a game can run a whole
TAS but isn't considered 'playable', but this is indisputably a serious bug.

I also changed the names of the output EXEs in build_nes.bat to match the ones
in the makefile.