Fix bad output on controller 2
It turns out that writing to controller port 1 actually strobes both
controllers, so not accounting for that was causing controller 2 to output
bad data.

I also sorted out the relationship between the APU and controller on register
$4017. It turns out that the APU only cares about writes, and the controller
only cares about reads, so there isn't really any conflict.

Finally, I changed it so controller output after reading all the buttons is
always 1, to make it act more like the real controller. Mostly this shouldn't
make a difference but the extra accuracy is nice.