Add structure for cycle-level CPU overhaul
I've started working on a process of moving over from the old
instruction-level code in parts, so that I can implement individual
instructions in the new system rather than trying to transition all at once,
which definitely wouldn't work very well. Only a few instructions are
implemented now, but the scaffolding is there to keep building on and finish
the rest of them.