New CPU core fully supported.
Everything works now, and the old CPU core has been entirely removed without
issue. I also spent some time trying to replicate the vblank read race
condition, with shakier results. Obviously, it's very sensitive to both CPU
and PPU accuracy, so I'll need to investigate it in greater depth.

The PPU has a few more minor tweaks to more closely imitate how FCEUX operates.
It now starts up on scanline 240 pixel 0, skipping the vblank on the immediate
next scanline 241. This appears to match FCEUX frame-for-frame. But it's
difficult to confirm because it's so deep in the emulator's internals.