Update vblank skip
It seems like I was overcomplicating things...FCEUX appears to simply skip
the first vblank, every time, no matter what. Why does it do that? It doesn't
appear to line up with the documentation I've read. I'm not sure at this point
if FCUEX implements true vblank skip at all, so maybe it doesn't really matter.
This appears to fix all the movies that are off-by-one-frame, so good riddance.

I also did a bit of cleanup on the CPU side of things, as all the ops don't
need to be set to do nothing by default anymore.