Fix array out of bounds issues
It turns out that when the PPU stored garbage data in the palette, the renderer
could try to access a palette value past the end of the array. I'm making sure
now that only the lower six bits of the pixel data are ever used, which should
keep it within the 64 existing palette entries.

I also fixed a potential issue where the vram address could increment outside
the PPU's address range.