Implement BRK instruction
The way NES DEV's documentation talks about the BRK instruction kinda implies
that it's not important for the NES, and indeed, it's not even tested for in
nestest or the like. So I concluded I didn't actually need to implement it.
After all, every game up till this point worked without it.

Bzzt, wrong. Dragon Warrior requires the BRK instruction to work, as it turns
out. So I went ahead and implemented it. It's probably not the cleanest
implementation, but I'm pretty burned out on CPU core stuff, so this will do.

I also removed a debug line from the MMC1 mapper.