Add support for large PRG ROMs in MMC1
Certain boards that use MMC1 use the CHR bank register in an unusual way.
They use bit 4 to switch between two 256 KB banks, to allow the cart to use
512 KB of PRG ROM. So this special behavior has to be detected and emulated as
well.

There are also some boards that use bits 2 and 3 of the CHR bank register
to bank switch for expanded PRG RAM. This isn't fully implemented, but I
tested ROMs known to use these boards, and they appear to work. It's possible
that they would run into issues further in, but I don't have time to test it
and see when I don't even really know where or when they might break. So I'm
leaving it as-is.

To support different PRG RAM sizes, I also moved the initialization of PRG RAM
into the mapper init code.