Add MMC3 mapper
The MMC3 mapper is now supported. This also required adding a number of
features to the CPU and PPU, including handling for IRQs and generalizing
interrupt handling a bit more.

Also implemented masking the leftmost tile, which isn't strictly needed,
but Super Mario 3 looks terrible without it. I also moved where sprite loading
happens, since it affects MMC3's scanline counter. There are still some issues
with the horizontal split in SMB3, but I think this is good enough to check in.