Fix MMC3 scanline clock not being generated on prerender scanline
MMC3's scanline counter depends on the way the PPU reads background tiles
and sprites. Previously, I had not been reading sprite data on the prerender
scanline, which caused the scanline counter to not be clocked on that scanline.
This has now been fixed, which should correct the IRQ timing for MMC3 games.